Added proper http error code
