+++ explorer/explorerroutes.go
@@ -32,7 +32,7 @@ func (exp *explorerUI) Home(w http.ResponseWriter, r *http.Request) {
 
 	if err != nil {
 		log.Errorf("Template execute failure: %v", err)
-		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things")
+		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things", false)
 		return
 	}
 	w.Header().Set("Content-Type", "text/html")
@@ -55,7 +55,7 @@ func (exp *explorerUI) Blocks(w http.ResponseWriter, r *http.Request) {
 	summaries := exp.blockData.GetExplorerBlocks(height, height-rows)
 	if summaries == nil {
 		log.Errorf("Unable to get blocks: height=%d&rows=%d", height, rows)
-		exp.ErrorPage(w, "Something went wrong...", "could not find those blocks")
+		exp.ErrorPage(w, "Something went wrong...", "could not find those blocks", true)
 		return
 	}
 
@@ -69,7 +69,7 @@ func (exp *explorerUI) Blocks(w http.ResponseWriter, r *http.Request) {
 
 	if err != nil {
 		log.Errorf("Template execute failure: %v", err)
-		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things")
+		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things", false)
 		return
 	}
 	w.Header().Set("Content-Type", "text/html")
@@ -83,7 +83,7 @@ func (exp *explorerUI) Block(w http.ResponseWriter, r *http.Request) {
 	data := exp.blockData.GetExplorerBlock(hash)
 	if data == nil {
 		log.Errorf("Unable to get block %s", hash)
-		exp.ErrorPage(w, "Something went wrong...", "could not find that block")
+		exp.ErrorPage(w, "Something went wrong...", "could not find that block", true)
 		return
 	}
 
@@ -97,7 +97,7 @@ func (exp *explorerUI) Block(w http.ResponseWriter, r *http.Request) {
 	str, err := templateExecToString(exp.templates[blockTemplateIndex], "block", pageData)
 	if err != nil {
 		log.Errorf("Template execute failure: %v", err)
-		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things")
+		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things", false)
 		return
 	}
 	w.Header().Set("Content-Type", "text/html")
@@ -110,13 +110,13 @@ func (exp *explorerUI) TxPage(w http.ResponseWriter, r *http.Request) {
 	hash, ok := r.Context().Value(ctxTxHash).(string)
 	if !ok {
 		log.Trace("txid not set")
-		exp.ErrorPage(w, "Something went wrong...", "there was no transaction requested")
+		exp.ErrorPage(w, "Something went wrong...", "there was no transaction requested", true)
 		return
 	}
 	tx := exp.blockData.GetExplorerTx(hash)
 	if tx == nil {
 		log.Errorf("Unable to get transaction %s", hash)
-		exp.ErrorPage(w, "Something went wrong...", "could not find that transaction")
+		exp.ErrorPage(w, "Something went wrong...", "could not find that transaction", true)
 		return
 	}
 	if !exp.liteMode {
@@ -125,7 +125,7 @@ func (exp *explorerUI) TxPage(w http.ResponseWriter, r *http.Request) {
 		spendingTxHashes, spendingTxVinInds, voutInds, err := exp.explorerSource.SpendingTransactions(hash)
 		if err != nil {
 			log.Errorf("Unable to retrieve spending transactions for %s: %v", hash, err)
-			exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things")
+			exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things", false)
 			return
 		}
 		for i, vout := range voutInds {
@@ -151,7 +151,7 @@ func (exp *explorerUI) TxPage(w http.ResponseWriter, r *http.Request) {
 	str, err := templateExecToString(exp.templates[txTemplateIndex], "tx", pageData)
 	if err != nil {
 		log.Errorf("Template execute failure: %v", err)
-		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things")
+		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things", false)
 		return
 	}
 	w.Header().Set("Content-Type", "text/html")
@@ -165,7 +165,7 @@ func (exp *explorerUI) AddressPage(w http.ResponseWriter, r *http.Request) {
 	address, ok := r.Context().Value(ctxAddress).(string)
 	if !ok {
 		log.Trace("address not set")
-		exp.ErrorPage(w, "Something went wrong...", "there seems to not be an address in this request")
+		exp.ErrorPage(w, "Something went wrong...", "there seems to not be an address in this request", true)
 		return
 	}
 
@@ -192,7 +192,7 @@ func (exp *explorerUI) AddressPage(w http.ResponseWriter, r *http.Request) {
 		addrData = exp.blockData.GetExplorerAddress(address, limitN, offsetAddrOuts)
 		if addrData == nil {
 			log.Errorf("Unable to get address %s", address)
-			exp.ErrorPage(w, "Something went wrong...", "could not find that address")
+			exp.ErrorPage(w, "Something went wrong...", "could not find that address", true)
 			return
 		}
 	} else {
@@ -201,7 +201,7 @@ func (exp *explorerUI) AddressPage(w http.ResponseWriter, r *http.Request) {
 			address, limitN, offsetAddrOuts)
 		if errH != nil {
 			log.Errorf("Unable to get address %s history: %v", address, errH)
-			exp.ErrorPage(w, "Something went wrong...", "could not find that address's history")
+			exp.ErrorPage(w, "Something went wrong...", "could not find that address's history", false)
 			return
 		}
 
@@ -209,7 +209,7 @@ func (exp *explorerUI) AddressPage(w http.ResponseWriter, r *http.Request) {
 		addrData = ReduceAddressHistory(addrHist)
 		if addrData == nil {
 			log.Debugf("empty address history (%s): n=%d&start=%d", address, limitN, offsetAddrOuts)
-			exp.ErrorPage(w, "Something went wrong...", "that address has no history")
+			exp.ErrorPage(w, "Something went wrong...", "that address has no history", true)
 			return
 		}
 		addrData.Limit, addrData.Offset = limitN, offsetAddrOuts
@@ -222,7 +222,7 @@ func (exp *explorerUI) AddressPage(w http.ResponseWriter, r *http.Request) {
 		err = exp.explorerSource.FillAddressTransactions(addrData)
 		if err != nil {
 			log.Errorf("Unable to fill address %s transactions: %v", address, err)
-			exp.ErrorPage(w, "Something went wrong...", "could not find transactions for that address")
+			exp.ErrorPage(w, "Something went wrong...", "could not find transactions for that address", false)
 			return
 		}
 	}
@@ -242,7 +242,7 @@ func (exp *explorerUI) AddressPage(w http.ResponseWriter, r *http.Request) {
 	str, err := templateExecToString(exp.templates[addressTemplateIndex], "address", pageData)
 	if err != nil {
 		log.Errorf("Template execute failure: %v", err)
-		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things")
+		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing... that usually fixes things", false)
 		return
 	}
 	w.Header().Set("Content-Type", "text/html")
@@ -254,7 +254,7 @@ func (exp *explorerUI) DecodeTxPage(w http.ResponseWriter, r *http.Request) {
 	str, err := templateExecToString(exp.templates[decodeTxTemplateIndex], "rawtx", nil)
 	if err != nil {
 		log.Errorf("Template execute failure: %v", err)
-		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing, that usually fixes things")
+		exp.ErrorPage(w, "Something went wrong...", "and it's not your fault, try refreshing, that usually fixes things", false)
 		return
 	}
 	w.Header().Set("Content-Type", "text/html")
@@ -308,7 +308,7 @@ func (exp *explorerUI) search(searchStr string) string {
 	return ""
 }
 
-func (exp *explorerUI) ErrorPage(w http.ResponseWriter, code string, message string) {
+func (exp *explorerUI) ErrorPage(w http.ResponseWriter, code string, message string, notFound bool) {
 	str, err := templateExecToString(exp.templates[errorTemplateIndex], "error", struct {
 		ErrorCode   string
 		ErrorString string
@@ -321,10 +321,14 @@ func (exp *explorerUI) ErrorPage(w http.ResponseWriter, code string, message str
 		str = "Something went very wrong if you can see this, try refreshing"
 	}
 	w.Header().Set("Content-Type", "text/html")
-	w.WriteHeader(http.StatusInternalServerError)
+	if notFound {
+		w.WriteHeader(http.StatusNotFound)
+	} else {
+		w.WriteHeader(http.StatusInternalServerError)
+	}
 	io.WriteString(w, str)
 }
 
 func (exp *explorerUI) NotFound(w http.ResponseWriter, r *http.Request) {
-	exp.ErrorPage(w, "Not found", "Cannot find page: "+r.URL.Path)
+	exp.ErrorPage(w, "Not found", "Cannot find page: "+r.URL.Path, true)
 }
