+++ views/extras.tmpl
@@ -168,6 +168,8 @@
             console.log("Got message: ", evt);
             if (evt !== "") {
                 window.location.href = evt;
+            } else {
+                //Search Error
             }
         })
         var updateBlockData = function (event) {
